requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0
Pillow==10.1.0
opencv-python==4.8.1.78
ultralytics==8.0.196
pyzbar==0.1.9
zxing-cpp==2.2.0
easyocr==1.7.0
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
numpy==1.24.3
gunicorn==21.2.0